error message return error code and trace info
