+++ client/option.go
@@ -1,0 +1,4 @@
+package client
+
+type ObkvOption struct {
+}
