+++ config/client_config_test.go
