--- error/code.go
@@ -1,1 +1,0 @@
-package error
