+++ error/error_code.go
@@ -1,0 +1,702 @@
+package error
+
+type ObErrorCode int32
+
+const (
+	ObSuccess                                ObErrorCode = 0
+	ObMaxErrorCode                           ObErrorCode = 10000
+	ObLastErrorCode                          ObErrorCode = -9018
+	ObErrSqlStart                            ObErrorCode = -5000
+	ObErrSqlEnd                              ObErrorCode = -5999
+	ObError                                  ObErrorCode = -4000
+	ObObjTypeError                           ObErrorCode = -4001
+	ObInvalidArgument                        ObErrorCode = -4002
+	ObArrayOutOfRange                        ObErrorCode = -4003
+	ObServerListenError                      ObErrorCode = -4004
+	ObInitTwice                              ObErrorCode = -4005
+	ObNotInit                                ObErrorCode = -4006
+	ObNotSupported                           ObErrorCode = -4007
+	ObIterEnd                                ObErrorCode = -4008
+	ObIoError                                ObErrorCode = -4009
+	ObErrorFuncVersion                       ObErrorCode = -4010
+	ObPacketNotSent                          ObErrorCode = -4011
+	ObTimeout                                ObErrorCode = -4012
+	ObAllocateMemoryFailed                   ObErrorCode = -4013
+	ObInnerStatError                         ObErrorCode = -4014
+	ObErrSys                                 ObErrorCode = -4015
+	ObErrUnexpected                          ObErrorCode = -4016
+	ObEntryExist                             ObErrorCode = -4017
+	ObEntryNotExist                          ObErrorCode = -4018
+	ObSizeOverflow                           ObErrorCode = -4019
+	ObRefNumNotZero                          ObErrorCode = -4020
+	ObConflictValue                          ObErrorCode = -4021
+	ObItemNotSetted                          ObErrorCode = -4022
+	ObEagain                                 ObErrorCode = -4023
+	ObBufNotEnough                           ObErrorCode = -4024
+	ObPartialFailed                          ObErrorCode = -4025
+	ObReadNothing                            ObErrorCode = -4026
+	ObFileNotExist                           ObErrorCode = -4027
+	ObDiscontinuousLog                       ObErrorCode = -4028
+	ObSchemaError                            ObErrorCode = -4029
+	ObTenantOutOfMem                         ObErrorCode = -4030
+	ObUnknownObj                             ObErrorCode = -4031
+	ObNoMonitorData                          ObErrorCode = -4032
+	ObSerializeError                         ObErrorCode = -4033
+	ObDeserializeError                       ObErrorCode = -4034
+	ObAioTimeout                             ObErrorCode = -4035
+	ObNeedRetry                              ObErrorCode = -4036
+	ObTooManySstable                         ObErrorCode = -4037
+	ObNotMaster                              ObErrorCode = -4038
+	ObDecryptFailed                          ObErrorCode = -4041
+	ObUserNotExist                           ObErrorCode = -4042
+	ObPasswordWrong                          ObErrorCode = -4043
+	ObSkeyVersionWrong                       ObErrorCode = -4044
+	ObNotRegistered                          ObErrorCode = -4048
+	ObWaitqueueTimeout                       ObErrorCode = -4049
+	ObNotTheObject                           ObErrorCode = -4050
+	ObAlreadyRegistered                      ObErrorCode = -4051
+	ObLastLogRuinned                         ObErrorCode = -4052
+	ObNoCsSelected                           ObErrorCode = -4053
+	ObNoTabletsCreated                       ObErrorCode = -4054
+	ObInvalidError                           ObErrorCode = -4055
+	ObDecimalOverflowWarn                    ObErrorCode = -4057
+	ObDecimalUnlegalError                    ObErrorCode = -4058
+	ObObjDivideError                         ObErrorCode = -4060
+	ObNotADecimal                            ObErrorCode = -4061
+	ObDecimalPrecisionNotEqual               ObErrorCode = -4062
+	ObEmptyRange                             ObErrorCode = -4063
+	ObSessionKilled                          ObErrorCode = -4064
+	ObLogNotSync                             ObErrorCode = -4065
+	ObDirNotExist                            ObErrorCode = -4066
+	ObSessionNotFound                        ObErrorCode = -4067
+	ObInvalidLog                             ObErrorCode = -4068
+	ObInvalidData                            ObErrorCode = -4070
+	ObAlreadyDone                            ObErrorCode = -4071
+	ObCanceled                               ObErrorCode = -4072
+	ObLogSrcChanged                          ObErrorCode = -4073
+	ObLogNotAlign                            ObErrorCode = -4074
+	ObLogMissing                             ObErrorCode = -4075
+	ObNeedWait                               ObErrorCode = -4076
+	ObNotImplement                           ObErrorCode = -4077
+	ObDivisionByZero                         ObErrorCode = -4078
+	ObExceedMemLimit                         ObErrorCode = -4080
+	ObResultUnknown                          ObErrorCode = -4081
+	ObNoResult                               ObErrorCode = -4084
+	ObQueueOverflow                          ObErrorCode = -4085
+	ObTermLagged                             ObErrorCode = -4097
+	ObTermNotMatch                           ObErrorCode = -4098
+	ObStartLogCursorInvalid                  ObErrorCode = -4099
+	ObLockNotMatch                           ObErrorCode = -4100
+	ObDeadLock                               ObErrorCode = -4101
+	ObPartialLog                             ObErrorCode = -4102
+	ObChecksumError                          ObErrorCode = -4103
+	ObInitFail                               ObErrorCode = -4104
+	ObNotEnoughStore                         ObErrorCode = -4106
+	ObBlockSwitched                          ObErrorCode = -4107
+	ObStateNotMatch                          ObErrorCode = -4109
+	ObReadZeroLog                            ObErrorCode = -4110
+	ObBlockNeedFreeze                        ObErrorCode = -4111
+	ObBlockFrozen                            ObErrorCode = -4112
+	ObInFatalState                           ObErrorCode = -4113
+	ObInStopState                            ObErrorCode = -4114
+	ObUpsMasterExists                        ObErrorCode = -4115
+	ObLogNotClear                            ObErrorCode = -4116
+	ObFileAlreadyExist                       ObErrorCode = -4117
+	ObUnknownPacket                          ObErrorCode = -4118
+	ObRpcPacketTooLong                       ObErrorCode = -4119
+	ObLogTooLarge                            ObErrorCode = -4120
+	ObRpcSendError                           ObErrorCode = -4121
+	ObRpcPostError                           ObErrorCode = -4122
+	ObLibeasyError                           ObErrorCode = -4123
+	ObConnectError                           ObErrorCode = -4124
+	ObNotFree                                ObErrorCode = -4125
+	ObInitSqlContextError                    ObErrorCode = -4126
+	ObSkipInvalidRow                         ObErrorCode = -4127
+	ObRpcPacketInvalid                       ObErrorCode = -4128
+	ObNoTablet                               ObErrorCode = -4133
+	ObSnapshotDiscarded                      ObErrorCode = -4138
+	ObDataNotUptodate                        ObErrorCode = -4139
+	ObRowModified                            ObErrorCode = -4142
+	ObVersionNotMatch                        ObErrorCode = -4143
+	ObBadAddress                             ObErrorCode = -4144
+	ObEnqueueFailed                          ObErrorCode = -4146
+	ObInvalidConfig                          ObErrorCode = -4147
+	ObStmtExpired                            ObErrorCode = -4149
+	ObErrMinValue                            ObErrorCode = -4150
+	ObErrMaxValue                            ObErrorCode = -4151
+	ObErrNullValue                           ObErrorCode = -4152
+	ObResourceOut                            ObErrorCode = -4153
+	ObErrSqlClient                           ObErrorCode = -4154
+	ObMetaTableWithoutUseTable               ObErrorCode = -4155
+	ObDiscardPacket                          ObErrorCode = -4156
+	ObOperateOverflow                        ObErrorCode = -4157
+	ObInvalidDateFormat                      ObErrorCode = -4158
+	ObPoolRegisteredFailed                   ObErrorCode = -4159
+	ObPoolUnregisteredFailed                 ObErrorCode = -4160
+	ObInvalidArgumentNum                     ObErrorCode = -4161
+	ObLeaseNotEnough                         ObErrorCode = -4162
+	ObLeaseNotMatch                          ObErrorCode = -4163
+	ObUpsSwitchNotHappen                     ObErrorCode = -4164
+	ObEmptyResult                            ObErrorCode = -4165
+	ObCacheNotHit                            ObErrorCode = -4166
+	ObNestedLoopNotSupport                   ObErrorCode = -4167
+	ObLogInvalidModId                        ObErrorCode = -4168
+	ObLogModuleUnknown                       ObErrorCode = -4169
+	ObLogLevelInvalid                        ObErrorCode = -4170
+	ObLogParserSyntaxErr                     ObErrorCode = -4171
+	ObIndexOutOfRange                        ObErrorCode = -4172
+	Ob_underflow                             ObErrorCode = -4173
+	ObUnknownConnection                      ObErrorCode = -4174
+	ObErrorOutOfRange                        ObErrorCode = -4175
+	ObCacheShrinkFailed                      ObErrorCode = -4176
+	ObOldSchemaVersion                       ObErrorCode = -4177
+	ObReleaseSchemaError                     ObErrorCode = -4178
+	ObOpNotAllow                             ObErrorCode = -4179
+	ObNoEmptyEntry                           ObErrorCode = -4180
+	ObErrAlreadyExists                       ObErrorCode = -4181
+	ObSearchNotFound                         ObErrorCode = -4182
+	ObBeyondTheRange                         ObErrorCode = -4183
+	ObCsOutofDiskSpace                       ObErrorCode = -4184
+	ObColumnGroupNotFound                    ObErrorCode = -4185
+	ObCsCompressLibError                     ObErrorCode = -4186
+	ObItemNotMatch                           ObErrorCode = -4187
+	ObSchedulerTaskCntMismatch               ObErrorCode = -4188
+	ObHashExist                              ObErrorCode = -4200
+	ObHashNotExist                           ObErrorCode = -4201
+	ObHashGetTimeout                         ObErrorCode = -4204
+	ObHashPlacementRetry                     ObErrorCode = -4205
+	ObHashFull                               ObErrorCode = -4206
+	ObPacketProcessed                        ObErrorCode = -4207
+	ObWaitNextTimeout                        ObErrorCode = -4208
+	ObLeaderNotExist                         ObErrorCode = -4209
+	ObPrepareMajorFreezeFailed               ObErrorCode = -4210
+	ObCommitMajorFreezeFailed                ObErrorCode = -4211
+	ObAbortMajorFreezeFailed                 ObErrorCode = -4212
+	ObMajorFreezeNotFinished                 ObErrorCode = -4213
+	ObPartitionNotLeader                     ObErrorCode = -4214
+	ObWaitMajorFreezeResponseTimeout         ObErrorCode = -4215
+	ObCurlError                              ObErrorCode = -4216
+	ObMajorFreezeNotAllow                    ObErrorCode = -4217
+	ObPrepareFreezeFailed                    ObErrorCode = -4218
+	ObInvalidDateValue                       ObErrorCode = -4219
+	ObInactiveSqlClient                      ObErrorCode = -4220
+	ObInactiveRpcProxy                       ObErrorCode = -4221
+	ObErvalWithMonth                         ObErrorCode = -4222
+	ObTooManyDatetimeParts                   ObErrorCode = -4223
+	ObDataOutOfRange                         ObErrorCode = -4224
+	ObPartitionNotExist                      ObErrorCode = -4225
+	ObErrTruncatedWrongValueForField         ObErrorCode = -4226
+	ObErrNoDefaultForField                   ObErrorCode = -4227
+	ObErrFieldSpecifiedTwice                 ObErrorCode = -4228
+	ObErrTooLongTableComment                 ObErrorCode = -4229
+	ObErrTooLongFieldComment                 ObErrorCode = -4230
+	ObErrTooLongIndexComment                 ObErrorCode = -4231
+	ObNotFollower                            ObErrorCode = -4232
+	ObErrOutOfLowerBound                     ObErrorCode = -4233
+	ObErrOutOfUpperBound                     ObErrorCode = -4234
+	ObBadNullError                           ObErrorCode = -4235
+	ObObconfigReturnError                    ObErrorCode = -4236
+	ObObconfigAppnameMismatch                ObErrorCode = -4237
+	ObErrViewSelectDerived                   ObErrorCode = -4238
+	ObCantMjPath                             ObErrorCode = -4239
+	ObErrNoJoinOrderGenerated                ObErrorCode = -4240
+	ObErrNoPathGenerated                     ObErrorCode = -4241
+	ObErrWaitRemoteSchemaRefresh             ObErrorCode = -4242
+	ObFileNotOpened                          ObErrorCode = -4243
+	ObTimerTaskHasScheduled                  ObErrorCode = -4244
+	ObTimerTaskHasNotScheduled               ObErrorCode = -4245
+	ObParseDebugSyncError                    ObErrorCode = -4246
+	ObUnknownDebugSyncPo                     ObErrorCode = -4247
+	ObErrErrupted                            ObErrorCode = -4248
+	ObErrDataTruncated                       ObErrorCode = -4249
+	ObNotRunning                             ObErrorCode = -4250
+	ObInvalidPartition                       ObErrorCode = -4251
+	ObErrTimeoutTruncated                    ObErrorCode = -4252
+	ObErrTooLongTenantComment                ObErrorCode = -4253
+	ObErrNetPacketTooLarge                   ObErrorCode = -4254
+	ObTraceDescNotExist                      ObErrorCode = -4255
+	ObErrNoDefault                           ObErrorCode = -4256
+	ObErrCompressDecompressData              ObErrorCode = -4257
+	ObErrIncorrectStringValue                ObErrorCode = -4258
+	ObErrDistributedNotSupported             ObErrorCode = -4259
+	ObIsChangingLeader                       ObErrorCode = -4260
+	ObDatetimeFunctionOverflow               ObErrorCode = -4261
+	ObErrDoubleTruncated                     ObErrorCode = -4262
+	ObMinorFreezeNotAllow                    ObErrorCode = -4263
+	ObLogOutofDiskSpace                      ObErrorCode = -4264
+	ObRpcConnectError                        ObErrorCode = -4265
+	ObMinorMergeNotAllow                     ObErrorCode = -4266
+	ObCacheInvalid                           ObErrorCode = -4267
+	ObReachServerDataCopyInConcurrencyLimit  ObErrorCode = -4268
+	ObWorkingPartitionExist                  ObErrorCode = -4269
+	ObWorkingPartitionNotExist               ObErrorCode = -4270
+	ObLibeasyReachMemLimit                   ObErrorCode = -4271
+	ObMissArgument                           ObErrorCode = -4272
+	ObCacheFreeBlockNotEnough                ObErrorCode = -4273
+	ObSyncWashMbTimeout                      ObErrorCode = -4274
+	ObNotAllowMigrateIn                      ObErrorCode = -4275
+	ObImportNotInServer                      ObErrorCode = -4505
+	ObConvertError                           ObErrorCode = -4507
+	ObBypassTimeout                          ObErrorCode = -4510
+	ObRsStateNotAllow                        ObErrorCode = -4512
+	ObNoReplicaValid                         ObErrorCode = -4515
+	ObNoNeedUpdate                           ObErrorCode = -4517
+	ObCacheTimeout                           ObErrorCode = -4518
+	ObIterStop                               ObErrorCode = -4519
+	ObZoneAlreadyMaster                      ObErrorCode = -4523
+	ObIpPortIsNotSlaveZone                   ObErrorCode = -4524
+	ObZoneIsNotSlave                         ObErrorCode = -4525
+	ObZoneIsNotMaster                        ObErrorCode = -4526
+	ObConfigNotSync                          ObErrorCode = -4527
+	ObIpPortIsNotZone                        ObErrorCode = -4528
+	ObMasterZoneNotExist                     ObErrorCode = -4529
+	ObZoneInfoNotExist                       ObErrorCode = -4530
+	ObGetZoneMasterUpsFailed                 ObErrorCode = -4531
+	ObMultipleMasterZonesExist               ObErrorCode = -4532
+	ObIndexingZoneInvalid                    ObErrorCode = -4533
+	ObRootTableRangeNotExist                 ObErrorCode = -4537
+	ObRootMigrateConcurrencyFull             ObErrorCode = -4538
+	ObRootMigrateInfoNotFound                ObErrorCode = -4539
+	ObNotDataLoadTable                       ObErrorCode = -4540
+	ObDataLoadTableDuplicated                ObErrorCode = -4541
+	ObRootTableIdExist                       ObErrorCode = -4542
+	ObIndexTimeout                           ObErrorCode = -4543
+	ObRootNotEgrated                         ObErrorCode = -4544
+	ObIndexIneligible                        ObErrorCode = -4545
+	ObRebalanceExecTimeout                   ObErrorCode = -4546
+	ObMergeNotStarted                        ObErrorCode = -4547
+	ObMergeAlreadyStarted                    ObErrorCode = -4548
+	ObRootserviceExist                       ObErrorCode = -4549
+	ObRsShutdown                             ObErrorCode = -4550
+	ObServerMigrateInDenied                  ObErrorCode = -4551
+	ObRebalanceTaskCantExec                  ObErrorCode = -4552
+	ObPartitionCntReachRootserverLimit       ObErrorCode = -4553
+	ObRebalanceTaskNotInProgress             ObErrorCode = -4554
+	ObDataSourceNotExist                     ObErrorCode = -4600
+	ObDataSourceTableNotExist                ObErrorCode = -4601
+	ObDataSourceRangeNotExist                ObErrorCode = -4602
+	ObDataSourceDataNotExist                 ObErrorCode = -4603
+	ObDataSourceSysError                     ObErrorCode = -4604
+	ObDataSourceTimeout                      ObErrorCode = -4605
+	ObDataSourceConcurrencyFull              ObErrorCode = -4606
+	ObDataSourceWrongUriFormat               ObErrorCode = -4607
+	ObSstableVersionUnequal                  ObErrorCode = -4608
+	ObUpsRenewLeaseNotAllowed                ObErrorCode = -4609
+	ObUpsCountOverLimit                      ObErrorCode = -4610
+	ObNoUpsMajority                          ObErrorCode = -4611
+	ObIndexCountReachTheLimit                ObErrorCode = -4613
+	ObTaskExpired                            ObErrorCode = -4614
+	ObTablegroupNotEmpty                     ObErrorCode = -4615
+	ObInvalidServerStatus                    ObErrorCode = -4620
+	ObWaitElecLeaderTimeout                  ObErrorCode = -4621
+	ObWaitAllRsOnlineTimeout                 ObErrorCode = -4622
+	ObAllReplicasOnMergeZone                 ObErrorCode = -4623
+	ObMachineResourceNotEnough               ObErrorCode = -4624
+	ObNotServerCanHoldSoftly                 ObErrorCode = -4625
+	ObResourcePoolAlreadyGranted             ObErrorCode = -4626
+	ObServerAlreadyDeleted                   ObErrorCode = -4628
+	ObServerNotDeleting                      ObErrorCode = -4629
+	ObServerNotInWhiteList                   ObErrorCode = -4630
+	ObServerZoneNotMatch                     ObErrorCode = -4631
+	ObOverZoneNumLimit                       ObErrorCode = -4632
+	ObZoneStatusNotMatch                     ObErrorCode = -4633
+	ObResourceUnitIsReferenced               ObErrorCode = -4634
+	ObDifferentPrimaryZone                   ObErrorCode = -4636
+	ObServerNotActive                        ObErrorCode = -4637
+	ObRsNotMaster                            ObErrorCode = -4638
+	ObCandidateListError                     ObErrorCode = -4639
+	ObPartitionZoneDuplicated                ObErrorCode = -4640
+	ObZoneDuplicated                         ObErrorCode = -4641
+	ObNotAllZoneActive                       ObErrorCode = -4642
+	ObPrimaryZoneNotInZoneList               ObErrorCode = -4643
+	ObReplicaNumNotMatch                     ObErrorCode = -4644
+	ObZoneListPoolListNotMatch               ObErrorCode = -4645
+	ObInvalidTenantName                      ObErrorCode = -4646
+	ObEmptyResourcePoolList                  ObErrorCode = -4647
+	ObResourceUnitNotExist                   ObErrorCode = -4648
+	ObResourceUnitExist                      ObErrorCode = -4649
+	ObResourcePoolNotExist                   ObErrorCode = -4650
+	ObResourcePoolExist                      ObErrorCode = -4651
+	ObWaitLeaderSwitchTimeout                ObErrorCode = -4652
+	ObLocationNotExist                       ObErrorCode = -4653
+	ObLocationLeaderNotExist                 ObErrorCode = -4654
+	ObZoneNotActive                          ObErrorCode = -4655
+	ObUnitNumOverServerCount                 ObErrorCode = -4656
+	ObPoolServerErsect                       ObErrorCode = -4657
+	ObNotSingleResourcePool                  ObErrorCode = -4658
+	ObInvalidResourceUnit                    ObErrorCode = -4659
+	ObStopServerInMultipleZones              ObErrorCode = -4660
+	ObSessionEntryExist                      ObErrorCode = -4661
+	ObGotSignalAborting                      ObErrorCode = -4662
+	ObServerNotAlive                         ObErrorCode = -4663
+	ObGetLocationTimeOut                     ObErrorCode = -4664
+	ObUnitIsMigrating                        ObErrorCode = -4665
+	ObClusterNoMatch                         ObErrorCode = -4666
+	ObCheckZoneMergeOrder                    ObErrorCode = -4667
+	ObErrZoneNotEmpty                        ObErrorCode = -4668
+	ObErrParserInit                          ObErrorCode = -5000
+	ObErrParseSql                            ObErrorCode = -5001
+	ObErrResolveSql                          ObErrorCode = -5002
+	ObErrGenPlan                             ObErrorCode = -5003
+	ObErrParserSyntax                        ObErrorCode = -5006
+	ObErrColumnSize                          ObErrorCode = -5007
+	ObErrColumnDuplicate                     ObErrorCode = -5008
+	ObErrOperatorUnknown                     ObErrorCode = -5010
+	ObErrStarDuplicate                       ObErrorCode = -5011
+	ObErrIllegalId                           ObErrorCode = -5012
+	ObErrIllegalValue                        ObErrorCode = -5014
+	ObErrColumnAmbiguous                     ObErrorCode = -5015
+	ObErrLogicalPlanFaild                    ObErrorCode = -5016
+	ObErrSchemaUnset                         ObErrorCode = -5017
+	ObErrIllegalName                         ObErrorCode = -5018
+	ObTableNotExist                          ObErrorCode = -5019
+	ObErrTableExist                          ObErrorCode = -5020
+	ObErrExprUnknown                         ObErrorCode = -5022
+	ObErrIllegalType                         ObErrorCode = -5023
+	ObErrPrimaryKeyDuplicate                 ObErrorCode = -5024
+	ObErrKeyNameDuplicate                    ObErrorCode = -5025
+	ObErrCreatetimeDuplicate                 ObErrorCode = -5026
+	ObErrModifytimeDuplicate                 ObErrorCode = -5027
+	ObErrIllegalIndex                        ObErrorCode = -5028
+	ObErrInvalidSchema                       ObErrorCode = -5029
+	ObErrInsertNullRowkey                    ObErrorCode = -5030
+	ObErrColumnNotFound                      ObErrorCode = -5031
+	ObErrDeleteNullRowkey                    ObErrorCode = -5032
+	ObErrUserEmpty                           ObErrorCode = -5034
+	ObErrUserNotExist                        ObErrorCode = -5035
+	ObErrNoPrivilege                         ObErrorCode = -5036
+	ObErrNoAvailablePrivilegeEntry           ObErrorCode = -5037
+	ObErrWrongPassword                       ObErrorCode = -5038
+	ObErrUserIsLocked                        ObErrorCode = -5039
+	ObErrUpdateRowkeyColumn                  ObErrorCode = -5040
+	ObErrUpdateJoinColumn                    ObErrorCode = -5041
+	ObErrInvalidColumnNum                    ObErrorCode = -5042
+	ObErrPrepareStmtNotFound                 ObErrorCode = -5043
+	ObErrSysVariableUnknown                  ObErrorCode = -5044
+	ObErrOlderPrivilegeVersion               ObErrorCode = -5046
+	ObErrLackOfRowkeyCol                     ObErrorCode = -5047
+	ObErrUserExist                           ObErrorCode = -5050
+	ObErrPasswordEmpty                       ObErrorCode = -5051
+	ObErrGrantPrivilegesToCreateTable        ObErrorCode = -5052
+	ObErrWrongDynamicParam                   ObErrorCode = -5053
+	ObErrParamSize                           ObErrorCode = -5054
+	ObErrFunctionUnknown                     ObErrorCode = -5055
+	ObErrCreatModifyTimeColumn               ObErrorCode = -5056
+	ObErrModifyPrimaryKey                    ObErrorCode = -5057
+	ObErrParamDuplicate                      ObErrorCode = -5058
+	ObErrTooManySessions                     ObErrorCode = -5059
+	ObErrTooManyPs                           ObErrorCode = -5061
+	ObErrHUnknown                            ObErrorCode = -5063
+	ObErrWhenUnsatisfied                     ObErrorCode = -5064
+	ObErrQueryErrupted                       ObErrorCode = -5065
+	ObErrSessionErrupted                     ObErrorCode = -5066
+	ObErrUnknownSessionId                    ObErrorCode = -5067
+	ObErrProtocolNotRecognize                ObErrorCode = -5068
+	ObErrWriteAuthError                      ObErrorCode = -5069
+	ObErrParseJoinInfo                       ObErrorCode = -5070
+	ObErrAlterIndexColumn                    ObErrorCode = -5071
+	ObErrModifyIndexTable                    ObErrorCode = -5072
+	ObErrIndexUnavailable                    ObErrorCode = -5073
+	ObErrNopValue                            ObErrorCode = -5074
+	ObErrPsTooManyParam                      ObErrorCode = -5080
+	ObErrReadOnly                            ObErrorCode = -5081
+	ObErrInvalidTypeForOp                    ObErrorCode = -5083
+	ObErrCastVarcharToBool                   ObErrorCode = -5084
+	ObErrCastVarcharToNumber                 ObErrorCode = -5085
+	ObErrCastVarcharToTime                   ObErrorCode = -5086
+	ObErrCastNumberOverflow                  ObErrorCode = -5087
+	ObEgerPrecisionOverflow                  ObErrorCode = -5088
+	ObDecimalPrecisionOverflow               ObErrorCode = -5089
+	ObSchemaNumberPrecisionOverflow          ObErrorCode = -5090
+	ObSchemaNumberScaleOverflow              ObErrorCode = -5091
+	ObErrIndexUnknown                        ObErrorCode = -5092
+	ObNumericOverflow                        ObErrorCode = -5093
+	ObErrTooManyJoinTables                   ObErrorCode = -5094
+	ObErrVarcharTooLong                      ObErrorCode = -5098
+	ObErrSysConfigUnknown                    ObErrorCode = -5099
+	ObErrLocalVariable                       ObErrorCode = -5100
+	ObErrGlobalVariable                      ObErrorCode = -5101
+	ObErrVariableIsReadonly                  ObErrorCode = -5102
+	ObErrIncorrectGlobalLocalVar             ObErrorCode = -5103
+	ObErrExpireInfoTooLong                   ObErrorCode = -5104
+	ObErrExpireCondTooLong                   ObErrorCode = -5105
+	ObInvalidArgumentForExtract              ObErrorCode = -5106
+	ObInvalidArgumentForIs                   ObErrorCode = -5107
+	ObInvalidArgumentForLength               ObErrorCode = -5108
+	ObInvalidArgumentForSubstr               ObErrorCode = -5109
+	ObInvalidArgumentForTimeToUsec           ObErrorCode = -5110
+	ObInvalidArgumentForUsecToTime           ObErrorCode = -5111
+	ObErrUserVariableUnknown                 ObErrorCode = -5112
+	ObIllegalUsageOfMergingFrozenTime        ObErrorCode = -5113
+	ObInvalidNumeric                         ObErrorCode = -5114
+	ObErrRegexpError                         ObErrorCode = -5115
+	ObSqlLogOpSetchildOverflow               ObErrorCode = -5116
+	ObSqlExplainFailed                       ObErrorCode = -5117
+	ObSqlOptCopyOpFailed                     ObErrorCode = -5118
+	ObSqlOptGenPlanFalied                    ObErrorCode = -5119
+	ObSqlOptCreateRawexprFailed              ObErrorCode = -5120
+	ObSqlOptJoinOrderFailed                  ObErrorCode = -5121
+	ObSqlOptError                            ObErrorCode = -5122
+	ObSqlResolverNoMemory                    ObErrorCode = -5130
+	ObSqlDmlOnly                             ObErrorCode = -5131
+	ObErrNoGrant                             ObErrorCode = -5133
+	ObErrNoDbSelected                        ObErrorCode = -5134
+	ObSqlPcOverflow                          ObErrorCode = -5135
+	ObSqlPcPlanDuplicate                     ObErrorCode = -5136
+	ObSqlPcPlanExpire                        ObErrorCode = -5137
+	ObSqlPcNotExist                          ObErrorCode = -5138
+	ObSqlParamsLimit                         ObErrorCode = -5139
+	ObSqlPcPlanSizeLimit                     ObErrorCode = -5140
+	ObErrUnknownCharset                      ObErrorCode = -5142
+	ObErrUnknownCollation                    ObErrorCode = -5143
+	ObErrCollationMismatch                   ObErrorCode = -5144
+	ObErrWrongValueForVar                    ObErrorCode = -5145
+	ObUnknownPartition                       ObErrorCode = -5146
+	ObPartitionNotMatch                      ObErrorCode = -5147
+	ObErPasswdLength                         ObErrorCode = -5148
+	ObErrInsertInnerJoinColumn               ObErrorCode = -5149
+	ObTenantNotInServer                      ObErrorCode = -5150
+	ObTablegroupNotExist                     ObErrorCode = -5151
+	ObSubqueryTooManyRow                     ObErrorCode = -5153
+	ObErrBadDatabase                         ObErrorCode = -5154
+	ObCannotUser                             ObErrorCode = -5155
+	ObTenantExist                            ObErrorCode = -5156
+	ObTenantNotExist                         ObErrorCode = -5157
+	ObDatabaseExist                          ObErrorCode = -5158
+	ObTablegroupExist                        ObErrorCode = -5159
+	ObErrInvalidTenantName                   ObErrorCode = -5160
+	ObEmptyTenant                            ObErrorCode = -5161
+	ObWrongDbName                            ObErrorCode = -5162
+	ObWrongTableName                         ObErrorCode = -5163
+	ObWrongColumnName                        ObErrorCode = -5164
+	ObErrColumnSpec                          ObErrorCode = -5165
+	ObErrDbDropExists                        ObErrorCode = -5166
+	ObErrDataTooLong                         ObErrorCode = -5167
+	ObErrWrongValueCountOnRow                ObErrorCode = -5168
+	ObErrCreateUserWithGrant                 ObErrorCode = -5169
+	ObErrNoDbPrivilege                       ObErrorCode = -5170
+	ObErrNoTablePrivilege                    ObErrorCode = -5171
+	ObInvalidOnUpdate                        ObErrorCode = -5172
+	ObInvalidDefault                         ObErrorCode = -5173
+	ObErrUpdateTableUsed                     ObErrorCode = -5174
+	ObErrCoulumnValueNotMatch                ObErrorCode = -5175
+	ObErrInvalidGroupFuncUse                 ObErrorCode = -5176
+	ObCantAggregate2collations               ObErrorCode = -5177
+	ObErrFieldTypeNotAllowedAsPartitionField ObErrorCode = -5178
+	ObErrTooLongIdent                        ObErrorCode = -5179
+	ObErrWrongTypeForVar                     ObErrorCode = -5180
+	ObWrongUserNameLength                    ObErrorCode = -5181
+	ObErrPrivUsage                           ObErrorCode = -5182
+	ObIllegalGrantForTable                   ObErrorCode = -5183
+	ObErrReachAutoincMax                     ObErrorCode = -5184
+	ObErrNoTablesUsed                        ObErrorCode = -5185
+	ObCantRemoveAllFields                    ObErrorCode = -5187
+	ObTooManyPartitionsError                 ObErrorCode = -5188
+	ObNoPartsError                           ObErrorCode = -5189
+	ObWrongSubKey                            ObErrorCode = -5190
+	ObKeyPart0                               ObErrorCode = -5191
+	ObErrUnknownTimeZone                     ObErrorCode = -5192
+	ObErrWrongAutoKey                        ObErrorCode = -5193
+	ObErrTooManyKeys                         ObErrorCode = -5194
+	ObErrTooManyRowkeyColumns                ObErrorCode = -5195
+	ObErrTooLongKeyLength                    ObErrorCode = -5196
+	ObErrTooManyColumns                      ObErrorCode = -5197
+	ObErrTooLongColumnLength                 ObErrorCode = -5198
+	ObErrTooBigRowsize                       ObErrorCode = -5199
+	ObErrUnknownTable                        ObErrorCode = -5200
+	ObErrBadTable                            ObErrorCode = -5201
+	ObErrTooBigScale                         ObErrorCode = -5202
+	ObErrTooBigPrecision                     ObErrorCode = -5203
+	ObErrMBiggerThanD                        ObErrorCode = -5204
+	ObErrTooBigDisplaywidth                  ObErrorCode = -5205
+	ObWrongGroupField                        ObErrorCode = -5206
+	ObNonUniqError                           ObErrorCode = -5207
+	ObErrNonuniqTable                        ObErrorCode = -5208
+	ObErrCantDropFieldOrKey                  ObErrorCode = -5209
+	ObErrMultiplePriKey                      ObErrorCode = -5210
+	ObErrKeyColumnDoesNotExits               ObErrorCode = -5211
+	ObErrAutoPartitionKey                    ObErrorCode = -5212
+	ObErrCantUseOptionHere                   ObErrorCode = -5213
+	ObErrWrongObject                         ObErrorCode = -5214
+	ObErrOnRename                            ObErrorCode = -5215
+	ObErrWrongKeyColumn                      ObErrorCode = -5216
+	ObErrBadFieldError                       ObErrorCode = -5217
+	ObErrWrongFieldWithGroup                 ObErrorCode = -5218
+	ObErrCantChangeTxCharacteristics         ObErrorCode = -5219
+	ObErrCantExecuteInReadOnlyTransaction    ObErrorCode = -5220
+	ObErrMixOfGroupFuncAndFields             ObErrorCode = -5221
+	ObErrTruncatedWrongValue                 ObErrorCode = -5222
+	ObErrWrongIdentName                      ObErrorCode = -5223
+	ObWrongNameForIndex                      ObErrorCode = -5224
+	ObIllegalReference                       ObErrorCode = -5225
+	ObReachMemoryLimit                       ObErrorCode = -5226
+	ObErrPasswordFormat                      ObErrorCode = -5227
+	ObErrNonUpdatableTable                   ObErrorCode = -5228
+	ObErrWarnDataOutOfRange                  ObErrorCode = -5229
+	ObErrWrongExprInPartitionFuncError       ObErrorCode = -5230
+	ObErrViewInvalid                         ObErrorCode = -5231
+	ObErrOptionPreventsStatement             ObErrorCode = -5233
+	ObErrDbReadOnly                          ObErrorCode = -5234
+	ObErrTableReadOnly                       ObErrorCode = -5235
+	ObErrLockOrActiveTransaction             ObErrorCode = -5236
+	ObErrSameNamePartitionField              ObErrorCode = -5237
+	ObErrTablenameNotAllowedHere             ObErrorCode = -5238
+	ObErrViewRecursive                       ObErrorCode = -5239
+	ObErrQualifier                           ObErrorCode = -5240
+	ObErrWrongValue                          ObErrorCode = -5241
+	ObErrViewWrongList                       ObErrorCode = -5242
+	ObSysVarsMaybeDiffVersion                ObErrorCode = -5243
+	ObErrAutoIncrementConflict               ObErrorCode = -5244
+	ObErrTaskSkipped                         ObErrorCode = -5245
+	ObErrNameBecomesEmpty                    ObErrorCode = -5246
+	ObErrRemovedSpaces                       ObErrorCode = -5247
+	ObWarnAddAutoincrementColumn             ObErrorCode = -5248
+	ObWarnChamgeNullAttribute                ObErrorCode = -5249
+	ObErrInvalidCharacterString              ObErrorCode = -5250
+	ObErrKillDenied                          ObErrorCode = -5251
+	ObErrColumnDefinitionAmbiguous           ObErrorCode = -5252
+	ObErrEmptyQuery                          ObErrorCode = -5253
+	ObErrCutValueGroupConcat                 ObErrorCode = -5254
+	ObErrFiledNotFoundPart                   ObErrorCode = -5255
+	ObErrPrimaryCantHaveNull                 ObErrorCode = -5256
+	ObErrPartitionFuncNotAllowedError        ObErrorCode = -5257
+	ObErrInvalidBlockSize                    ObErrorCode = -5258
+	ObErrUnknownStorageEngine                ObErrorCode = -5259
+	ObErrTenantIsLocked                      ObErrorCode = -5260
+	ObEerUniqueKeyNeedAllFieldsInPf          ObErrorCode = -5261
+	ObErrAggregateOrderForUnion              ObErrorCode = -5263
+	ObErrOutlineExist                        ObErrorCode = -5264
+	ObOutlineNotExist                        ObErrorCode = -5265
+	ObWarnOptionBelowLimit                   ObErrorCode = -5266
+	ObInvalidOutline                         ObErrorCode = -5267
+	ObReachMaxConcurrentNum                  ObErrorCode = -5268
+	ObErrOperationOnRecycleObject            ObErrorCode = -5269
+	ObErrObjectNotInRecyclebin               ObErrorCode = -5270
+	ObErrConCountError                       ObErrorCode = -5271
+	ObErrOutlineContentExist                 ObErrorCode = -5272
+	ObErrOutlineMaxConcurrentExist           ObErrorCode = -5273
+	ObErrValuesIsNot_typeError               ObErrorCode = -5274
+	ObErrWrongTypeColumnValueError           ObErrorCode = -5275
+	ObErrPartitionColumnListError            ObErrorCode = -5276
+	ObErrTooManyValuesError                  ObErrorCode = -5277
+	ObErrPartitionFunctionIsNotAllowed       ObErrorCode = -5278
+	ObErrPartitionErvalError                 ObErrorCode = -5279
+	ObErrSameNamePartition                   ObErrorCode = -5280
+	ObErrRangeNotIncreasingError             ObErrorCode = -5281
+	ObErrParsePartitionRange                 ObErrorCode = -5282
+	ObErrUniqueKeyNeedAllFieldsInPf          ObErrorCode = -5283
+	ObNoPartitionForGivenValue               ObErrorCode = -5284
+	ObEerNullInValuesLessThan                ObErrorCode = -5285
+	ObErrPartitionConstDomainError           ObErrorCode = -5286
+	ObErrTooManyPartitionFuncFields          ObErrorCode = -5287
+	ObErrBadFtColumn                         ObErrorCode = -5288
+	ObErrKeyDoesNotExists                    ObErrorCode = -5289
+	ObNonDefaultValueForGeneratedColumn      ObErrorCode = -5290
+	ObErrBadCtxcatColumn                     ObErrorCode = -5291
+	ObErrUnsupportedActionOnGeneratedColumn  ObErrorCode = -5292
+	ObErrDependentByGeneratedColumn          ObErrorCode = -5293
+	ObErrTooManyRows                         ObErrorCode = -5294
+	ObWrongFieldTerminators                  ObErrorCode = -5295
+	ObNoReadableReplica                      ObErrorCode = -5296
+	ObErrPartitionMgmtOnNonpartitioned       ObErrorCode = -5302
+	ObErrDropPartitionNonExistent            ObErrorCode = -5303
+	ObErrPartitionMgmtOnTwopartTable         ObErrorCode = -5304
+	ObErrOnlyOnRangeListPartition            ObErrorCode = -5305
+	ObErrDropLastPartition                   ObErrorCode = -5306
+	ObErrSchedulerThreadNotEnough            ObErrorCode = -5307
+	ObErrIgnoreUserHostName                  ObErrorCode = -5308
+	ObIgnoreSqlInRestore                     ObErrorCode = -5309
+	ObErrUnexpectedTzTransition              ObErrorCode = -5310
+	ObErrInvalidColumnId                     ObErrorCode = -5311
+	ObTransactionSetViolation                ObErrorCode = -6001
+	ObTransRollbacked                        ObErrorCode = -6002
+	ObErrExclusiveLockConflict               ObErrorCode = -6003
+	ObErrSharedLockConflict                  ObErrorCode = -6004
+	ObTryLockRowConflict                     ObErrorCode = -6005
+	ObClockOutOfOrder                        ObErrorCode = -6201
+	ObMaskSetNoNode                          ObErrorCode = -6203
+	ObTransHasDecided                        ObErrorCode = -6204
+	ObTransInvalidState                      ObErrorCode = -6205
+	ObTransStateNotChange                    ObErrorCode = -6206
+	ObTransProtocolError                     ObErrorCode = -6207
+	ObTransInvalidMessage                    ObErrorCode = -6208
+	ObTransInvalidMessageType                ObErrorCode = -6209
+	ObTransTimeout                           ObErrorCode = -6210
+	ObTransKilled                            ObErrorCode = -6211
+	ObTransStmtTimeout                       ObErrorCode = -6212
+	ObTransCtxNotExist                       ObErrorCode = -6213
+	ObPartitionIsFrozen                      ObErrorCode = -6214
+	ObPartitionIsNotFrozen                   ObErrorCode = -6215
+	ObTransInvalidLogType                    ObErrorCode = -6219
+	ObTransSqlSequenceIllegal                ObErrorCode = -6220
+	ObTransCannotBeKilled                    ObErrorCode = -6221
+	ObTransStateUnknown                      ObErrorCode = -6222
+	ObTransIsExiting                         ObErrorCode = -6223
+	ObTransNeedRollback                      ObErrorCode = -6224
+	ObTransUnknown                           ObErrorCode = -6225
+	ObErrReadOnlyTransaction                 ObErrorCode = -6226
+	ObPartitionIsNotStopped                  ObErrorCode = -6227
+	ObPartitionIsStopped                     ObErrorCode = -6228
+	ObPartitionIsBlocked                     ObErrorCode = -6229
+	ObTransRpcTimeout                        ObErrorCode = -6230
+	ObReplicaNotReadable                     ObErrorCode = -6231
+	ObLogIdNotFound                          ObErrorCode = -6301
+	ObLsrThreadStopped                       ObErrorCode = -6302
+	ObNoLog                                  ObErrorCode = -6303
+	ObLogIdRangeError                        ObErrorCode = -6304
+	ObLogIterEnough                          ObErrorCode = -6305
+	ObClogInvalidAck                         ObErrorCode = -6306
+	ObClogCacheInvalid                       ObErrorCode = -6307
+	ObExtHandleUnfinish                      ObErrorCode = -6308
+	ObCursorNotExist                         ObErrorCode = -6309
+	ObStreamNotExist                         ObErrorCode = -6310
+	ObStreamBusy                             ObErrorCode = -6311
+	ObFileRecycled                           ObErrorCode = -6312
+	ObReplayEagainTooMuchTime                ObErrorCode = -6313
+	ObMemberChangeFailed                     ObErrorCode = -6314
+	ObElectionWarnLogbufFull                 ObErrorCode = -7000
+	ObElectionWarnLogbufEmpty                ObErrorCode = -7001
+	ObElectionWarnNotRunning                 ObErrorCode = -7002
+	ObElectionWarnIsRunning                  ObErrorCode = -7003
+	ObElectionWarnNotReachMajority           ObErrorCode = -7004
+	ObElectionWarnInvalidServer              ObErrorCode = -7005
+	ObElectionWarnInvalidLeader              ObErrorCode = -7006
+	ObElectionWarnLeaderLeaseExpired         ObErrorCode = -7007
+	ObElectionWarnInvalidMessage             ObErrorCode = -7010
+	ObElectionWarnMessageNotIme              ObErrorCode = -7011
+	ObElectionWarnNotCandidate               ObErrorCode = -7012
+	ObElectionWarnNotCandidateOrVoter        ObErrorCode = -7013
+	ObElectionWarnProtocolError              ObErrorCode = -7014
+	ObElectionWarnRuntimeOutOfRange          ObErrorCode = -7015
+	ObElectionWarnLastOperationNotDone       ObErrorCode = -7021
+	ObElectionWarnCurrentServerNotLeader     ObErrorCode = -7022
+	ObElectionWarnNoPrepareMessage           ObErrorCode = -7024
+	ObElectionErrorMultiPrepareMessage       ObErrorCode = -7025
+	ObElectionNotExist                       ObErrorCode = -7026
+	ObElectionMgrIsRunning                   ObErrorCode = -7027
+	ObElectionWarnNoMajorityPrepareMessage   ObErrorCode = -7029
+	ObElectionAsyncLogWarnInit               ObErrorCode = -7030
+	ObElectionWaitLeaderMessage              ObErrorCode = -7031
+	ObServerIsInit                           ObErrorCode = -8001
+	ObServerIsStopping                       ObErrorCode = -8002
+	ObPacketChecksumError                    ObErrorCode = -8003
+	ObPacketClusterIdNotMatch                ObErrorCode = -8004
+	ObUriError                               ObErrorCode = -9001
+	ObFinalMd5Error                          ObErrorCode = -9002
+	ObOssError                               ObErrorCode = -9003
+	ObInitMd5Error                           ObErrorCode = -9004
+	ObOutOfElement                           ObErrorCode = -9005
+	ObUpdateMd5Error                         ObErrorCode = -9006
+	ObFileLengthInvalid                      ObErrorCode = -9007
+	ObNotReadAllData                         ObErrorCode = -9008
+	ObBuildMd5Error                          ObErrorCode = -9009
+	ObMd5NotMatch                            ObErrorCode = -9010
+	ObOssFileNotExist                        ObErrorCode = -9011
+	ObOssDataVersionNotMatched               ObErrorCode = -9012
+	ObOssWriteError                          ObErrorCode = -9013
+	ObRestoreInProgress                      ObErrorCode = -9014
+	ObAgentInitingBackupCountError           ObErrorCode = -9015
+	ObClusterNameNotEqual                    ObErrorCode = -9016
+	ObRsListInvaild                          ObErrorCode = -9017
+	ObAgentHasFailedTask                     ObErrorCode = -9018
+)
