--- obkv/obkv.go
@@ -1,1 +1,0 @@
-package obkv
