--- protocol/responsecodes.go
@@ -1,702 +1,0 @@
-package protocol
-
-type ResponseCode int32
-
-const (
-	ObSuccess                                ResponseCode = 0
-	ObMaxErrorCode                           ResponseCode = 10000
-	ObLastErrorCode                          ResponseCode = -9018
-	ObErrSqlStart                            ResponseCode = -5000
-	ObErrSqlEnd                              ResponseCode = -5999
-	ObError                                  ResponseCode = -4000
-	ObObjTypeError                           ResponseCode = -4001
-	ObInvalidArgument                        ResponseCode = -4002
-	ObArrayOutOfRange                        ResponseCode = -4003
-	ObServerListenError                      ResponseCode = -4004
-	ObInitTwice                              ResponseCode = -4005
-	ObNotInit                                ResponseCode = -4006
-	ObNotSupported                           ResponseCode = -4007
-	ObIterEnd                                ResponseCode = -4008
-	ObIoError                                ResponseCode = -4009
-	ObErrorFuncVersion                       ResponseCode = -4010
-	ObPacketNotSent                          ResponseCode = -4011
-	ObTimeout                                ResponseCode = -4012
-	ObAllocateMemoryFailed                   ResponseCode = -4013
-	ObInnerStatError                         ResponseCode = -4014
-	ObErrSys                                 ResponseCode = -4015
-	ObErrUnexpected                          ResponseCode = -4016
-	ObEntryExist                             ResponseCode = -4017
-	ObEntryNotExist                          ResponseCode = -4018
-	ObSizeOverflow                           ResponseCode = -4019
-	ObRefNumNotZero                          ResponseCode = -4020
-	ObConflictValue                          ResponseCode = -4021
-	ObItemNotSetted                          ResponseCode = -4022
-	ObEagain                                 ResponseCode = -4023
-	ObBufNotEnough                           ResponseCode = -4024
-	ObPartialFailed                          ResponseCode = -4025
-	ObReadNothing                            ResponseCode = -4026
-	ObFileNotExist                           ResponseCode = -4027
-	ObDiscontinuousLog                       ResponseCode = -4028
-	ObSchemaError                            ResponseCode = -4029
-	ObTenantOutOfMem                         ResponseCode = -4030
-	ObUnknownObj                             ResponseCode = -4031
-	ObNoMonitorData                          ResponseCode = -4032
-	ObSerializeError                         ResponseCode = -4033
-	ObDeserializeError                       ResponseCode = -4034
-	ObAioTimeout                             ResponseCode = -4035
-	ObNeedRetry                              ResponseCode = -4036
-	ObTooManySstable                         ResponseCode = -4037
-	ObNotMaster                              ResponseCode = -4038
-	ObDecryptFailed                          ResponseCode = -4041
-	ObUserNotExist                           ResponseCode = -4042
-	ObPasswordWrong                          ResponseCode = -4043
-	ObSkeyVersionWrong                       ResponseCode = -4044
-	ObNotRegistered                          ResponseCode = -4048
-	ObWaitqueueTimeout                       ResponseCode = -4049
-	ObNotTheObject                           ResponseCode = -4050
-	ObAlreadyRegistered                      ResponseCode = -4051
-	ObLastLogRuinned                         ResponseCode = -4052
-	ObNoCsSelected                           ResponseCode = -4053
-	ObNoTabletsCreated                       ResponseCode = -4054
-	ObInvalidError                           ResponseCode = -4055
-	ObDecimalOverflowWarn                    ResponseCode = -4057
-	ObDecimalUnlegalError                    ResponseCode = -4058
-	ObObjDivideError                         ResponseCode = -4060
-	ObNotADecimal                            ResponseCode = -4061
-	ObDecimalPrecisionNotEqual               ResponseCode = -4062
-	ObEmptyRange                             ResponseCode = -4063
-	ObSessionKilled                          ResponseCode = -4064
-	ObLogNotSync                             ResponseCode = -4065
-	ObDirNotExist                            ResponseCode = -4066
-	ObSessionNotFound                        ResponseCode = -4067
-	ObInvalidLog                             ResponseCode = -4068
-	ObInvalidData                            ResponseCode = -4070
-	ObAlreadyDone                            ResponseCode = -4071
-	ObCanceled                               ResponseCode = -4072
-	ObLogSrcChanged                          ResponseCode = -4073
-	ObLogNotAlign                            ResponseCode = -4074
-	ObLogMissing                             ResponseCode = -4075
-	ObNeedWait                               ResponseCode = -4076
-	ObNotImplement                           ResponseCode = -4077
-	ObDivisionByZero                         ResponseCode = -4078
-	ObExceedMemLimit                         ResponseCode = -4080
-	ObResultUnknown                          ResponseCode = -4081
-	ObNoResult                               ResponseCode = -4084
-	ObQueueOverflow                          ResponseCode = -4085
-	ObTermLagged                             ResponseCode = -4097
-	ObTermNotMatch                           ResponseCode = -4098
-	ObStartLogCursorInvalid                  ResponseCode = -4099
-	ObLockNotMatch                           ResponseCode = -4100
-	ObDeadLock                               ResponseCode = -4101
-	ObPartialLog                             ResponseCode = -4102
-	ObChecksumError                          ResponseCode = -4103
-	ObInitFail                               ResponseCode = -4104
-	ObNotEnoughStore                         ResponseCode = -4106
-	ObBlockSwitched                          ResponseCode = -4107
-	ObStateNotMatch                          ResponseCode = -4109
-	ObReadZeroLog                            ResponseCode = -4110
-	ObBlockNeedFreeze                        ResponseCode = -4111
-	ObBlockFrozen                            ResponseCode = -4112
-	ObInFatalState                           ResponseCode = -4113
-	ObInStopState                            ResponseCode = -4114
-	ObUpsMasterExists                        ResponseCode = -4115
-	ObLogNotClear                            ResponseCode = -4116
-	ObFileAlreadyExist                       ResponseCode = -4117
-	ObUnknownPacket                          ResponseCode = -4118
-	ObRpcPacketTooLong                       ResponseCode = -4119
-	ObLogTooLarge                            ResponseCode = -4120
-	ObRpcSendError                           ResponseCode = -4121
-	ObRpcPostError                           ResponseCode = -4122
-	ObLibeasyError                           ResponseCode = -4123
-	ObConnectError                           ResponseCode = -4124
-	ObNotFree                                ResponseCode = -4125
-	ObInitSqlContextError                    ResponseCode = -4126
-	ObSkipInvalidRow                         ResponseCode = -4127
-	ObRpcPacketInvalid                       ResponseCode = -4128
-	ObNoTablet                               ResponseCode = -4133
-	ObSnapshotDiscarded                      ResponseCode = -4138
-	ObDataNotUptodate                        ResponseCode = -4139
-	ObRowModified                            ResponseCode = -4142
-	ObVersionNotMatch                        ResponseCode = -4143
-	ObBadAddress                             ResponseCode = -4144
-	ObEnqueueFailed                          ResponseCode = -4146
-	ObInvalidConfig                          ResponseCode = -4147
-	ObStmtExpired                            ResponseCode = -4149
-	ObErrMinValue                            ResponseCode = -4150
-	ObErrMaxValue                            ResponseCode = -4151
-	ObErrNullValue                           ResponseCode = -4152
-	ObResourceOut                            ResponseCode = -4153
-	ObErrSqlClient                           ResponseCode = -4154
-	ObMetaTableWithoutUseTable               ResponseCode = -4155
-	ObDiscardPacket                          ResponseCode = -4156
-	ObOperateOverflow                        ResponseCode = -4157
-	ObInvalidDateFormat                      ResponseCode = -4158
-	ObPoolRegisteredFailed                   ResponseCode = -4159
-	ObPoolUnregisteredFailed                 ResponseCode = -4160
-	ObInvalidArgumentNum                     ResponseCode = -4161
-	ObLeaseNotEnough                         ResponseCode = -4162
-	ObLeaseNotMatch                          ResponseCode = -4163
-	ObUpsSwitchNotHappen                     ResponseCode = -4164
-	ObEmptyResult                            ResponseCode = -4165
-	ObCacheNotHit                            ResponseCode = -4166
-	ObNestedLoopNotSupport                   ResponseCode = -4167
-	ObLogInvalidModId                        ResponseCode = -4168
-	ObLogModuleUnknown                       ResponseCode = -4169
-	ObLogLevelInvalid                        ResponseCode = -4170
-	ObLogParserSyntaxErr                     ResponseCode = -4171
-	ObIndexOutOfRange                        ResponseCode = -4172
-	Ob_underflow                             ResponseCode = -4173
-	ObUnknownConnection                      ResponseCode = -4174
-	ObErrorOutOfRange                        ResponseCode = -4175
-	ObCacheShrinkFailed                      ResponseCode = -4176
-	ObOldSchemaVersion                       ResponseCode = -4177
-	ObReleaseSchemaError                     ResponseCode = -4178
-	ObOpNotAllow                             ResponseCode = -4179
-	ObNoEmptyEntry                           ResponseCode = -4180
-	ObErrAlreadyExists                       ResponseCode = -4181
-	ObSearchNotFound                         ResponseCode = -4182
-	ObBeyondTheRange                         ResponseCode = -4183
-	ObCsOutofDiskSpace                       ResponseCode = -4184
-	ObColumnGroupNotFound                    ResponseCode = -4185
-	ObCsCompressLibError                     ResponseCode = -4186
-	ObItemNotMatch                           ResponseCode = -4187
-	ObSchedulerTaskCntMismatch               ResponseCode = -4188
-	ObHashExist                              ResponseCode = -4200
-	ObHashNotExist                           ResponseCode = -4201
-	ObHashGetTimeout                         ResponseCode = -4204
-	ObHashPlacementRetry                     ResponseCode = -4205
-	ObHashFull                               ResponseCode = -4206
-	ObPacketProcessed                        ResponseCode = -4207
-	ObWaitNextTimeout                        ResponseCode = -4208
-	ObLeaderNotExist                         ResponseCode = -4209
-	ObPrepareMajorFreezeFailed               ResponseCode = -4210
-	ObCommitMajorFreezeFailed                ResponseCode = -4211
-	ObAbortMajorFreezeFailed                 ResponseCode = -4212
-	ObMajorFreezeNotFinished                 ResponseCode = -4213
-	ObPartitionNotLeader                     ResponseCode = -4214
-	ObWaitMajorFreezeResponseTimeout         ResponseCode = -4215
-	ObCurlError                              ResponseCode = -4216
-	ObMajorFreezeNotAllow                    ResponseCode = -4217
-	ObPrepareFreezeFailed                    ResponseCode = -4218
-	ObInvalidDateValue                       ResponseCode = -4219
-	ObInactiveSqlClient                      ResponseCode = -4220
-	ObInactiveRpcProxy                       ResponseCode = -4221
-	ObErvalWithMonth                         ResponseCode = -4222
-	ObTooManyDatetimeParts                   ResponseCode = -4223
-	ObDataOutOfRange                         ResponseCode = -4224
-	ObPartitionNotExist                      ResponseCode = -4225
-	ObErrTruncatedWrongValueForField         ResponseCode = -4226
-	ObErrNoDefaultForField                   ResponseCode = -4227
-	ObErrFieldSpecifiedTwice                 ResponseCode = -4228
-	ObErrTooLongTableComment                 ResponseCode = -4229
-	ObErrTooLongFieldComment                 ResponseCode = -4230
-	ObErrTooLongIndexComment                 ResponseCode = -4231
-	ObNotFollower                            ResponseCode = -4232
-	ObErrOutOfLowerBound                     ResponseCode = -4233
-	ObErrOutOfUpperBound                     ResponseCode = -4234
-	ObBadNullError                           ResponseCode = -4235
-	ObObconfigReturnError                    ResponseCode = -4236
-	ObObconfigAppnameMismatch                ResponseCode = -4237
-	ObErrViewSelectDerived                   ResponseCode = -4238
-	ObCantMjPath                             ResponseCode = -4239
-	ObErrNoJoinOrderGenerated                ResponseCode = -4240
-	ObErrNoPathGenerated                     ResponseCode = -4241
-	ObErrWaitRemoteSchemaRefresh             ResponseCode = -4242
-	ObFileNotOpened                          ResponseCode = -4243
-	ObTimerTaskHasScheduled                  ResponseCode = -4244
-	ObTimerTaskHasNotScheduled               ResponseCode = -4245
-	ObParseDebugSyncError                    ResponseCode = -4246
-	ObUnknownDebugSyncPo                     ResponseCode = -4247
-	ObErrErrupted                            ResponseCode = -4248
-	ObErrDataTruncated                       ResponseCode = -4249
-	ObNotRunning                             ResponseCode = -4250
-	ObInvalidPartition                       ResponseCode = -4251
-	ObErrTimeoutTruncated                    ResponseCode = -4252
-	ObErrTooLongTenantComment                ResponseCode = -4253
-	ObErrNetPacketTooLarge                   ResponseCode = -4254
-	ObTraceDescNotExist                      ResponseCode = -4255
-	ObErrNoDefault                           ResponseCode = -4256
-	ObErrCompressDecompressData              ResponseCode = -4257
-	ObErrIncorrectStringValue                ResponseCode = -4258
-	ObErrDistributedNotSupported             ResponseCode = -4259
-	ObIsChangingLeader                       ResponseCode = -4260
-	ObDatetimeFunctionOverflow               ResponseCode = -4261
-	ObErrDoubleTruncated                     ResponseCode = -4262
-	ObMinorFreezeNotAllow                    ResponseCode = -4263
-	ObLogOutofDiskSpace                      ResponseCode = -4264
-	ObRpcConnectError                        ResponseCode = -4265
-	ObMinorMergeNotAllow                     ResponseCode = -4266
-	ObCacheInvalid                           ResponseCode = -4267
-	ObReachServerDataCopyInConcurrencyLimit  ResponseCode = -4268
-	ObWorkingPartitionExist                  ResponseCode = -4269
-	ObWorkingPartitionNotExist               ResponseCode = -4270
-	ObLibeasyReachMemLimit                   ResponseCode = -4271
-	ObMissArgument                           ResponseCode = -4272
-	ObCacheFreeBlockNotEnough                ResponseCode = -4273
-	ObSyncWashMbTimeout                      ResponseCode = -4274
-	ObNotAllowMigrateIn                      ResponseCode = -4275
-	ObImportNotInServer                      ResponseCode = -4505
-	ObConvertError                           ResponseCode = -4507
-	ObBypassTimeout                          ResponseCode = -4510
-	ObRsStateNotAllow                        ResponseCode = -4512
-	ObNoReplicaValid                         ResponseCode = -4515
-	ObNoNeedUpdate                           ResponseCode = -4517
-	ObCacheTimeout                           ResponseCode = -4518
-	ObIterStop                               ResponseCode = -4519
-	ObZoneAlreadyMaster                      ResponseCode = -4523
-	ObIpPortIsNotSlaveZone                   ResponseCode = -4524
-	ObZoneIsNotSlave                         ResponseCode = -4525
-	ObZoneIsNotMaster                        ResponseCode = -4526
-	ObConfigNotSync                          ResponseCode = -4527
-	ObIpPortIsNotZone                        ResponseCode = -4528
-	ObMasterZoneNotExist                     ResponseCode = -4529
-	ObZoneInfoNotExist                       ResponseCode = -4530
-	ObGetZoneMasterUpsFailed                 ResponseCode = -4531
-	ObMultipleMasterZonesExist               ResponseCode = -4532
-	ObIndexingZoneInvalid                    ResponseCode = -4533
-	ObRootTableRangeNotExist                 ResponseCode = -4537
-	ObRootMigrateConcurrencyFull             ResponseCode = -4538
-	ObRootMigrateInfoNotFound                ResponseCode = -4539
-	ObNotDataLoadTable                       ResponseCode = -4540
-	ObDataLoadTableDuplicated                ResponseCode = -4541
-	ObRootTableIdExist                       ResponseCode = -4542
-	ObIndexTimeout                           ResponseCode = -4543
-	ObRootNotEgrated                         ResponseCode = -4544
-	ObIndexIneligible                        ResponseCode = -4545
-	ObRebalanceExecTimeout                   ResponseCode = -4546
-	ObMergeNotStarted                        ResponseCode = -4547
-	ObMergeAlreadyStarted                    ResponseCode = -4548
-	ObRootserviceExist                       ResponseCode = -4549
-	ObRsShutdown                             ResponseCode = -4550
-	ObServerMigrateInDenied                  ResponseCode = -4551
-	ObRebalanceTaskCantExec                  ResponseCode = -4552
-	ObPartitionCntReachRootserverLimit       ResponseCode = -4553
-	ObRebalanceTaskNotInProgress             ResponseCode = -4554
-	ObDataSourceNotExist                     ResponseCode = -4600
-	ObDataSourceTableNotExist                ResponseCode = -4601
-	ObDataSourceRangeNotExist                ResponseCode = -4602
-	ObDataSourceDataNotExist                 ResponseCode = -4603
-	ObDataSourceSysError                     ResponseCode = -4604
-	ObDataSourceTimeout                      ResponseCode = -4605
-	ObDataSourceConcurrencyFull              ResponseCode = -4606
-	ObDataSourceWrongUriFormat               ResponseCode = -4607
-	ObSstableVersionUnequal                  ResponseCode = -4608
-	ObUpsRenewLeaseNotAllowed                ResponseCode = -4609
-	ObUpsCountOverLimit                      ResponseCode = -4610
-	ObNoUpsMajority                          ResponseCode = -4611
-	ObIndexCountReachTheLimit                ResponseCode = -4613
-	ObTaskExpired                            ResponseCode = -4614
-	ObTablegroupNotEmpty                     ResponseCode = -4615
-	ObInvalidServerStatus                    ResponseCode = -4620
-	ObWaitElecLeaderTimeout                  ResponseCode = -4621
-	ObWaitAllRsOnlineTimeout                 ResponseCode = -4622
-	ObAllReplicasOnMergeZone                 ResponseCode = -4623
-	ObMachineResourceNotEnough               ResponseCode = -4624
-	ObNotServerCanHoldSoftly                 ResponseCode = -4625
-	ObResourcePoolAlreadyGranted             ResponseCode = -4626
-	ObServerAlreadyDeleted                   ResponseCode = -4628
-	ObServerNotDeleting                      ResponseCode = -4629
-	ObServerNotInWhiteList                   ResponseCode = -4630
-	ObServerZoneNotMatch                     ResponseCode = -4631
-	ObOverZoneNumLimit                       ResponseCode = -4632
-	ObZoneStatusNotMatch                     ResponseCode = -4633
-	ObResourceUnitIsReferenced               ResponseCode = -4634
-	ObDifferentPrimaryZone                   ResponseCode = -4636
-	ObServerNotActive                        ResponseCode = -4637
-	ObRsNotMaster                            ResponseCode = -4638
-	ObCandidateListError                     ResponseCode = -4639
-	ObPartitionZoneDuplicated                ResponseCode = -4640
-	ObZoneDuplicated                         ResponseCode = -4641
-	ObNotAllZoneActive                       ResponseCode = -4642
-	ObPrimaryZoneNotInZoneList               ResponseCode = -4643
-	ObReplicaNumNotMatch                     ResponseCode = -4644
-	ObZoneListPoolListNotMatch               ResponseCode = -4645
-	ObInvalidTenantName                      ResponseCode = -4646
-	ObEmptyResourcePoolList                  ResponseCode = -4647
-	ObResourceUnitNotExist                   ResponseCode = -4648
-	ObResourceUnitExist                      ResponseCode = -4649
-	ObResourcePoolNotExist                   ResponseCode = -4650
-	ObResourcePoolExist                      ResponseCode = -4651
-	ObWaitLeaderSwitchTimeout                ResponseCode = -4652
-	ObLocationNotExist                       ResponseCode = -4653
-	ObLocationLeaderNotExist                 ResponseCode = -4654
-	ObZoneNotActive                          ResponseCode = -4655
-	ObUnitNumOverServerCount                 ResponseCode = -4656
-	ObPoolServerErsect                       ResponseCode = -4657
-	ObNotSingleResourcePool                  ResponseCode = -4658
-	ObInvalidResourceUnit                    ResponseCode = -4659
-	ObStopServerInMultipleZones              ResponseCode = -4660
-	ObSessionEntryExist                      ResponseCode = -4661
-	ObGotSignalAborting                      ResponseCode = -4662
-	ObServerNotAlive                         ResponseCode = -4663
-	ObGetLocationTimeOut                     ResponseCode = -4664
-	ObUnitIsMigrating                        ResponseCode = -4665
-	ObClusterNoMatch                         ResponseCode = -4666
-	ObCheckZoneMergeOrder                    ResponseCode = -4667
-	ObErrZoneNotEmpty                        ResponseCode = -4668
-	ObErrParserInit                          ResponseCode = -5000
-	ObErrParseSql                            ResponseCode = -5001
-	ObErrResolveSql                          ResponseCode = -5002
-	ObErrGenPlan                             ResponseCode = -5003
-	ObErrParserSyntax                        ResponseCode = -5006
-	ObErrColumnSize                          ResponseCode = -5007
-	ObErrColumnDuplicate                     ResponseCode = -5008
-	ObErrOperatorUnknown                     ResponseCode = -5010
-	ObErrStarDuplicate                       ResponseCode = -5011
-	ObErrIllegalId                           ResponseCode = -5012
-	ObErrIllegalValue                        ResponseCode = -5014
-	ObErrColumnAmbiguous                     ResponseCode = -5015
-	ObErrLogicalPlanFaild                    ResponseCode = -5016
-	ObErrSchemaUnset                         ResponseCode = -5017
-	ObErrIllegalName                         ResponseCode = -5018
-	ObTableNotExist                          ResponseCode = -5019
-	ObErrTableExist                          ResponseCode = -5020
-	ObErrExprUnknown                         ResponseCode = -5022
-	ObErrIllegalType                         ResponseCode = -5023
-	ObErrPrimaryKeyDuplicate                 ResponseCode = -5024
-	ObErrKeyNameDuplicate                    ResponseCode = -5025
-	ObErrCreatetimeDuplicate                 ResponseCode = -5026
-	ObErrModifytimeDuplicate                 ResponseCode = -5027
-	ObErrIllegalIndex                        ResponseCode = -5028
-	ObErrInvalidSchema                       ResponseCode = -5029
-	ObErrInsertNullRowkey                    ResponseCode = -5030
-	ObErrColumnNotFound                      ResponseCode = -5031
-	ObErrDeleteNullRowkey                    ResponseCode = -5032
-	ObErrUserEmpty                           ResponseCode = -5034
-	ObErrUserNotExist                        ResponseCode = -5035
-	ObErrNoPrivilege                         ResponseCode = -5036
-	ObErrNoAvailablePrivilegeEntry           ResponseCode = -5037
-	ObErrWrongPassword                       ResponseCode = -5038
-	ObErrUserIsLocked                        ResponseCode = -5039
-	ObErrUpdateRowkeyColumn                  ResponseCode = -5040
-	ObErrUpdateJoinColumn                    ResponseCode = -5041
-	ObErrInvalidColumnNum                    ResponseCode = -5042
-	ObErrPrepareStmtNotFound                 ResponseCode = -5043
-	ObErrSysVariableUnknown                  ResponseCode = -5044
-	ObErrOlderPrivilegeVersion               ResponseCode = -5046
-	ObErrLackOfRowkeyCol                     ResponseCode = -5047
-	ObErrUserExist                           ResponseCode = -5050
-	ObErrPasswordEmpty                       ResponseCode = -5051
-	ObErrGrantPrivilegesToCreateTable        ResponseCode = -5052
-	ObErrWrongDynamicParam                   ResponseCode = -5053
-	ObErrParamSize                           ResponseCode = -5054
-	ObErrFunctionUnknown                     ResponseCode = -5055
-	ObErrCreatModifyTimeColumn               ResponseCode = -5056
-	ObErrModifyPrimaryKey                    ResponseCode = -5057
-	ObErrParamDuplicate                      ResponseCode = -5058
-	ObErrTooManySessions                     ResponseCode = -5059
-	ObErrTooManyPs                           ResponseCode = -5061
-	ObErrHUnknown                            ResponseCode = -5063
-	ObErrWhenUnsatisfied                     ResponseCode = -5064
-	ObErrQueryErrupted                       ResponseCode = -5065
-	ObErrSessionErrupted                     ResponseCode = -5066
-	ObErrUnknownSessionId                    ResponseCode = -5067
-	ObErrProtocolNotRecognize                ResponseCode = -5068
-	ObErrWriteAuthError                      ResponseCode = -5069
-	ObErrParseJoinInfo                       ResponseCode = -5070
-	ObErrAlterIndexColumn                    ResponseCode = -5071
-	ObErrModifyIndexTable                    ResponseCode = -5072
-	ObErrIndexUnavailable                    ResponseCode = -5073
-	ObErrNopValue                            ResponseCode = -5074
-	ObErrPsTooManyParam                      ResponseCode = -5080
-	ObErrReadOnly                            ResponseCode = -5081
-	ObErrInvalidTypeForOp                    ResponseCode = -5083
-	ObErrCastVarcharToBool                   ResponseCode = -5084
-	ObErrCastVarcharToNumber                 ResponseCode = -5085
-	ObErrCastVarcharToTime                   ResponseCode = -5086
-	ObErrCastNumberOverflow                  ResponseCode = -5087
-	ObEgerPrecisionOverflow                  ResponseCode = -5088
-	ObDecimalPrecisionOverflow               ResponseCode = -5089
-	ObSchemaNumberPrecisionOverflow          ResponseCode = -5090
-	ObSchemaNumberScaleOverflow              ResponseCode = -5091
-	ObErrIndexUnknown                        ResponseCode = -5092
-	ObNumericOverflow                        ResponseCode = -5093
-	ObErrTooManyJoinTables                   ResponseCode = -5094
-	ObErrVarcharTooLong                      ResponseCode = -5098
-	ObErrSysConfigUnknown                    ResponseCode = -5099
-	ObErrLocalVariable                       ResponseCode = -5100
-	ObErrGlobalVariable                      ResponseCode = -5101
-	ObErrVariableIsReadonly                  ResponseCode = -5102
-	ObErrIncorrectGlobalLocalVar             ResponseCode = -5103
-	ObErrExpireInfoTooLong                   ResponseCode = -5104
-	ObErrExpireCondTooLong                   ResponseCode = -5105
-	ObInvalidArgumentForExtract              ResponseCode = -5106
-	ObInvalidArgumentForIs                   ResponseCode = -5107
-	ObInvalidArgumentForLength               ResponseCode = -5108
-	ObInvalidArgumentForSubstr               ResponseCode = -5109
-	ObInvalidArgumentForTimeToUsec           ResponseCode = -5110
-	ObInvalidArgumentForUsecToTime           ResponseCode = -5111
-	ObErrUserVariableUnknown                 ResponseCode = -5112
-	ObIllegalUsageOfMergingFrozenTime        ResponseCode = -5113
-	ObInvalidNumeric                         ResponseCode = -5114
-	ObErrRegexpError                         ResponseCode = -5115
-	ObSqlLogOpSetchildOverflow               ResponseCode = -5116
-	ObSqlExplainFailed                       ResponseCode = -5117
-	ObSqlOptCopyOpFailed                     ResponseCode = -5118
-	ObSqlOptGenPlanFalied                    ResponseCode = -5119
-	ObSqlOptCreateRawexprFailed              ResponseCode = -5120
-	ObSqlOptJoinOrderFailed                  ResponseCode = -5121
-	ObSqlOptError                            ResponseCode = -5122
-	ObSqlResolverNoMemory                    ResponseCode = -5130
-	ObSqlDmlOnly                             ResponseCode = -5131
-	ObErrNoGrant                             ResponseCode = -5133
-	ObErrNoDbSelected                        ResponseCode = -5134
-	ObSqlPcOverflow                          ResponseCode = -5135
-	ObSqlPcPlanDuplicate                     ResponseCode = -5136
-	ObSqlPcPlanExpire                        ResponseCode = -5137
-	ObSqlPcNotExist                          ResponseCode = -5138
-	ObSqlParamsLimit                         ResponseCode = -5139
-	ObSqlPcPlanSizeLimit                     ResponseCode = -5140
-	ObErrUnknownCharset                      ResponseCode = -5142
-	ObErrUnknownCollation                    ResponseCode = -5143
-	ObErrCollationMismatch                   ResponseCode = -5144
-	ObErrWrongValueForVar                    ResponseCode = -5145
-	ObUnknownPartition                       ResponseCode = -5146
-	ObPartitionNotMatch                      ResponseCode = -5147
-	ObErPasswdLength                         ResponseCode = -5148
-	ObErrInsertInnerJoinColumn               ResponseCode = -5149
-	ObTenantNotInServer                      ResponseCode = -5150
-	ObTablegroupNotExist                     ResponseCode = -5151
-	ObSubqueryTooManyRow                     ResponseCode = -5153
-	ObErrBadDatabase                         ResponseCode = -5154
-	ObCannotUser                             ResponseCode = -5155
-	ObTenantExist                            ResponseCode = -5156
-	ObTenantNotExist                         ResponseCode = -5157
-	ObDatabaseExist                          ResponseCode = -5158
-	ObTablegroupExist                        ResponseCode = -5159
-	ObErrInvalidTenantName                   ResponseCode = -5160
-	ObEmptyTenant                            ResponseCode = -5161
-	ObWrongDbName                            ResponseCode = -5162
-	ObWrongTableName                         ResponseCode = -5163
-	ObWrongColumnName                        ResponseCode = -5164
-	ObErrColumnSpec                          ResponseCode = -5165
-	ObErrDbDropExists                        ResponseCode = -5166
-	ObErrDataTooLong                         ResponseCode = -5167
-	ObErrWrongValueCountOnRow                ResponseCode = -5168
-	ObErrCreateUserWithGrant                 ResponseCode = -5169
-	ObErrNoDbPrivilege                       ResponseCode = -5170
-	ObErrNoTablePrivilege                    ResponseCode = -5171
-	ObInvalidOnUpdate                        ResponseCode = -5172
-	ObInvalidDefault                         ResponseCode = -5173
-	ObErrUpdateTableUsed                     ResponseCode = -5174
-	ObErrCoulumnValueNotMatch                ResponseCode = -5175
-	ObErrInvalidGroupFuncUse                 ResponseCode = -5176
-	ObCantAggregate2collations               ResponseCode = -5177
-	ObErrFieldTypeNotAllowedAsPartitionField ResponseCode = -5178
-	ObErrTooLongIdent                        ResponseCode = -5179
-	ObErrWrongTypeForVar                     ResponseCode = -5180
-	ObWrongUserNameLength                    ResponseCode = -5181
-	ObErrPrivUsage                           ResponseCode = -5182
-	ObIllegalGrantForTable                   ResponseCode = -5183
-	ObErrReachAutoincMax                     ResponseCode = -5184
-	ObErrNoTablesUsed                        ResponseCode = -5185
-	ObCantRemoveAllFields                    ResponseCode = -5187
-	ObTooManyPartitionsError                 ResponseCode = -5188
-	ObNoPartsError                           ResponseCode = -5189
-	ObWrongSubKey                            ResponseCode = -5190
-	ObKeyPart0                               ResponseCode = -5191
-	ObErrUnknownTimeZone                     ResponseCode = -5192
-	ObErrWrongAutoKey                        ResponseCode = -5193
-	ObErrTooManyKeys                         ResponseCode = -5194
-	ObErrTooManyRowkeyColumns                ResponseCode = -5195
-	ObErrTooLongKeyLength                    ResponseCode = -5196
-	ObErrTooManyColumns                      ResponseCode = -5197
-	ObErrTooLongColumnLength                 ResponseCode = -5198
-	ObErrTooBigRowsize                       ResponseCode = -5199
-	ObErrUnknownTable                        ResponseCode = -5200
-	ObErrBadTable                            ResponseCode = -5201
-	ObErrTooBigScale                         ResponseCode = -5202
-	ObErrTooBigPrecision                     ResponseCode = -5203
-	ObErrMBiggerThanD                        ResponseCode = -5204
-	ObErrTooBigDisplaywidth                  ResponseCode = -5205
-	ObWrongGroupField                        ResponseCode = -5206
-	ObNonUniqError                           ResponseCode = -5207
-	ObErrNonuniqTable                        ResponseCode = -5208
-	ObErrCantDropFieldOrKey                  ResponseCode = -5209
-	ObErrMultiplePriKey                      ResponseCode = -5210
-	ObErrKeyColumnDoesNotExits               ResponseCode = -5211
-	ObErrAutoPartitionKey                    ResponseCode = -5212
-	ObErrCantUseOptionHere                   ResponseCode = -5213
-	ObErrWrongObject                         ResponseCode = -5214
-	ObErrOnRename                            ResponseCode = -5215
-	ObErrWrongKeyColumn                      ResponseCode = -5216
-	ObErrBadFieldError                       ResponseCode = -5217
-	ObErrWrongFieldWithGroup                 ResponseCode = -5218
-	ObErrCantChangeTxCharacteristics         ResponseCode = -5219
-	ObErrCantExecuteInReadOnlyTransaction    ResponseCode = -5220
-	ObErrMixOfGroupFuncAndFields             ResponseCode = -5221
-	ObErrTruncatedWrongValue                 ResponseCode = -5222
-	ObErrWrongIdentName                      ResponseCode = -5223
-	ObWrongNameForIndex                      ResponseCode = -5224
-	ObIllegalReference                       ResponseCode = -5225
-	ObReachMemoryLimit                       ResponseCode = -5226
-	ObErrPasswordFormat                      ResponseCode = -5227
-	ObErrNonUpdatableTable                   ResponseCode = -5228
-	ObErrWarnDataOutOfRange                  ResponseCode = -5229
-	ObErrWrongExprInPartitionFuncError       ResponseCode = -5230
-	ObErrViewInvalid                         ResponseCode = -5231
-	ObErrOptionPreventsStatement             ResponseCode = -5233
-	ObErrDbReadOnly                          ResponseCode = -5234
-	ObErrTableReadOnly                       ResponseCode = -5235
-	ObErrLockOrActiveTransaction             ResponseCode = -5236
-	ObErrSameNamePartitionField              ResponseCode = -5237
-	ObErrTablenameNotAllowedHere             ResponseCode = -5238
-	ObErrViewRecursive                       ResponseCode = -5239
-	ObErrQualifier                           ResponseCode = -5240
-	ObErrWrongValue                          ResponseCode = -5241
-	ObErrViewWrongList                       ResponseCode = -5242
-	ObSysVarsMaybeDiffVersion                ResponseCode = -5243
-	ObErrAutoIncrementConflict               ResponseCode = -5244
-	ObErrTaskSkipped                         ResponseCode = -5245
-	ObErrNameBecomesEmpty                    ResponseCode = -5246
-	ObErrRemovedSpaces                       ResponseCode = -5247
-	ObWarnAddAutoincrementColumn             ResponseCode = -5248
-	ObWarnChamgeNullAttribute                ResponseCode = -5249
-	ObErrInvalidCharacterString              ResponseCode = -5250
-	ObErrKillDenied                          ResponseCode = -5251
-	ObErrColumnDefinitionAmbiguous           ResponseCode = -5252
-	ObErrEmptyQuery                          ResponseCode = -5253
-	ObErrCutValueGroupConcat                 ResponseCode = -5254
-	ObErrFiledNotFoundPart                   ResponseCode = -5255
-	ObErrPrimaryCantHaveNull                 ResponseCode = -5256
-	ObErrPartitionFuncNotAllowedError        ResponseCode = -5257
-	ObErrInvalidBlockSize                    ResponseCode = -5258
-	ObErrUnknownStorageEngine                ResponseCode = -5259
-	ObErrTenantIsLocked                      ResponseCode = -5260
-	ObEerUniqueKeyNeedAllFieldsInPf          ResponseCode = -5261
-	ObErrAggregateOrderForUnion              ResponseCode = -5263
-	ObErrOutlineExist                        ResponseCode = -5264
-	ObOutlineNotExist                        ResponseCode = -5265
-	ObWarnOptionBelowLimit                   ResponseCode = -5266
-	ObInvalidOutline                         ResponseCode = -5267
-	ObReachMaxConcurrentNum                  ResponseCode = -5268
-	ObErrOperationOnRecycleObject            ResponseCode = -5269
-	ObErrObjectNotInRecyclebin               ResponseCode = -5270
-	ObErrConCountError                       ResponseCode = -5271
-	ObErrOutlineContentExist                 ResponseCode = -5272
-	ObErrOutlineMaxConcurrentExist           ResponseCode = -5273
-	ObErrValuesIsNot_typeError               ResponseCode = -5274
-	ObErrWrongTypeColumnValueError           ResponseCode = -5275
-	ObErrPartitionColumnListError            ResponseCode = -5276
-	ObErrTooManyValuesError                  ResponseCode = -5277
-	ObErrPartitionFunctionIsNotAllowed       ResponseCode = -5278
-	ObErrPartitionErvalError                 ResponseCode = -5279
-	ObErrSameNamePartition                   ResponseCode = -5280
-	ObErrRangeNotIncreasingError             ResponseCode = -5281
-	ObErrParsePartitionRange                 ResponseCode = -5282
-	ObErrUniqueKeyNeedAllFieldsInPf          ResponseCode = -5283
-	ObNoPartitionForGivenValue               ResponseCode = -5284
-	ObEerNullInValuesLessThan                ResponseCode = -5285
-	ObErrPartitionConstDomainError           ResponseCode = -5286
-	ObErrTooManyPartitionFuncFields          ResponseCode = -5287
-	ObErrBadFtColumn                         ResponseCode = -5288
-	ObErrKeyDoesNotExists                    ResponseCode = -5289
-	ObNonDefaultValueForGeneratedColumn      ResponseCode = -5290
-	ObErrBadCtxcatColumn                     ResponseCode = -5291
-	ObErrUnsupportedActionOnGeneratedColumn  ResponseCode = -5292
-	ObErrDependentByGeneratedColumn          ResponseCode = -5293
-	ObErrTooManyRows                         ResponseCode = -5294
-	ObWrongFieldTerminators                  ResponseCode = -5295
-	ObNoReadableReplica                      ResponseCode = -5296
-	ObErrPartitionMgmtOnNonpartitioned       ResponseCode = -5302
-	ObErrDropPartitionNonExistent            ResponseCode = -5303
-	ObErrPartitionMgmtOnTwopartTable         ResponseCode = -5304
-	ObErrOnlyOnRangeListPartition            ResponseCode = -5305
-	ObErrDropLastPartition                   ResponseCode = -5306
-	ObErrSchedulerThreadNotEnough            ResponseCode = -5307
-	ObErrIgnoreUserHostName                  ResponseCode = -5308
-	ObIgnoreSqlInRestore                     ResponseCode = -5309
-	ObErrUnexpectedTzTransition              ResponseCode = -5310
-	ObErrInvalidColumnId                     ResponseCode = -5311
-	ObTransactionSetViolation                ResponseCode = -6001
-	ObTransRollbacked                        ResponseCode = -6002
-	ObErrExclusiveLockConflict               ResponseCode = -6003
-	ObErrSharedLockConflict                  ResponseCode = -6004
-	ObTryLockRowConflict                     ResponseCode = -6005
-	ObClockOutOfOrder                        ResponseCode = -6201
-	ObMaskSetNoNode                          ResponseCode = -6203
-	ObTransHasDecided                        ResponseCode = -6204
-	ObTransInvalidState                      ResponseCode = -6205
-	ObTransStateNotChange                    ResponseCode = -6206
-	ObTransProtocolError                     ResponseCode = -6207
-	ObTransInvalidMessage                    ResponseCode = -6208
-	ObTransInvalidMessageType                ResponseCode = -6209
-	ObTransTimeout                           ResponseCode = -6210
-	ObTransKilled                            ResponseCode = -6211
-	ObTransStmtTimeout                       ResponseCode = -6212
-	ObTransCtxNotExist                       ResponseCode = -6213
-	ObPartitionIsFrozen                      ResponseCode = -6214
-	ObPartitionIsNotFrozen                   ResponseCode = -6215
-	ObTransInvalidLogType                    ResponseCode = -6219
-	ObTransSqlSequenceIllegal                ResponseCode = -6220
-	ObTransCannotBeKilled                    ResponseCode = -6221
-	ObTransStateUnknown                      ResponseCode = -6222
-	ObTransIsExiting                         ResponseCode = -6223
-	ObTransNeedRollback                      ResponseCode = -6224
-	ObTransUnknown                           ResponseCode = -6225
-	ObErrReadOnlyTransaction                 ResponseCode = -6226
-	ObPartitionIsNotStopped                  ResponseCode = -6227
-	ObPartitionIsStopped                     ResponseCode = -6228
-	ObPartitionIsBlocked                     ResponseCode = -6229
-	ObTransRpcTimeout                        ResponseCode = -6230
-	ObReplicaNotReadable                     ResponseCode = -6231
-	ObLogIdNotFound                          ResponseCode = -6301
-	ObLsrThreadStopped                       ResponseCode = -6302
-	ObNoLog                                  ResponseCode = -6303
-	ObLogIdRangeError                        ResponseCode = -6304
-	ObLogIterEnough                          ResponseCode = -6305
-	ObClogInvalidAck                         ResponseCode = -6306
-	ObClogCacheInvalid                       ResponseCode = -6307
-	ObExtHandleUnfinish                      ResponseCode = -6308
-	ObCursorNotExist                         ResponseCode = -6309
-	ObStreamNotExist                         ResponseCode = -6310
-	ObStreamBusy                             ResponseCode = -6311
-	ObFileRecycled                           ResponseCode = -6312
-	ObReplayEagainTooMuchTime                ResponseCode = -6313
-	ObMemberChangeFailed                     ResponseCode = -6314
-	ObElectionWarnLogbufFull                 ResponseCode = -7000
-	ObElectionWarnLogbufEmpty                ResponseCode = -7001
-	ObElectionWarnNotRunning                 ResponseCode = -7002
-	ObElectionWarnIsRunning                  ResponseCode = -7003
-	ObElectionWarnNotReachMajority           ResponseCode = -7004
-	ObElectionWarnInvalidServer              ResponseCode = -7005
-	ObElectionWarnInvalidLeader              ResponseCode = -7006
-	ObElectionWarnLeaderLeaseExpired         ResponseCode = -7007
-	ObElectionWarnInvalidMessage             ResponseCode = -7010
-	ObElectionWarnMessageNotIme              ResponseCode = -7011
-	ObElectionWarnNotCandidate               ResponseCode = -7012
-	ObElectionWarnNotCandidateOrVoter        ResponseCode = -7013
-	ObElectionWarnProtocolError              ResponseCode = -7014
-	ObElectionWarnRuntimeOutOfRange          ResponseCode = -7015
-	ObElectionWarnLastOperationNotDone       ResponseCode = -7021
-	ObElectionWarnCurrentServerNotLeader     ResponseCode = -7022
-	ObElectionWarnNoPrepareMessage           ResponseCode = -7024
-	ObElectionErrorMultiPrepareMessage       ResponseCode = -7025
-	ObElectionNotExist                       ResponseCode = -7026
-	ObElectionMgrIsRunning                   ResponseCode = -7027
-	ObElectionWarnNoMajorityPrepareMessage   ResponseCode = -7029
-	ObElectionAsyncLogWarnInit               ResponseCode = -7030
-	ObElectionWaitLeaderMessage              ResponseCode = -7031
-	ObServerIsInit                           ResponseCode = -8001
-	ObServerIsStopping                       ResponseCode = -8002
-	ObPacketChecksumError                    ResponseCode = -8003
-	ObPacketClusterIdNotMatch                ResponseCode = -8004
-	ObUriError                               ResponseCode = -9001
-	ObFinalMd5Error                          ResponseCode = -9002
-	ObOssError                               ResponseCode = -9003
-	ObInitMd5Error                           ResponseCode = -9004
-	ObOutOfElement                           ResponseCode = -9005
-	ObUpdateMd5Error                         ResponseCode = -9006
-	ObFileLengthInvalid                      ResponseCode = -9007
-	ObNotReadAllData                         ResponseCode = -9008
-	ObBuildMd5Error                          ResponseCode = -9009
-	ObMd5NotMatch                            ResponseCode = -9010
-	ObOssFileNotExist                        ResponseCode = -9011
-	ObOssDataVersionNotMatched               ResponseCode = -9012
-	ObOssWriteError                          ResponseCode = -9013
-	ObRestoreInProgress                      ResponseCode = -9014
-	ObAgentInitingBackupCountError           ResponseCode = -9015
-	ObClusterNameNotEqual                    ResponseCode = -9016
-	ObRsListInvaild                          ResponseCode = -9017
-	ObAgentHasFailedTask                     ResponseCode = -9018
-)
