+++ route/partition_location_entry.go
@@ -1,0 +1,35 @@
+package route
+
+import "strconv"
+
+type ObPartLocationEntry struct {
+	partLocations map[int64]*ObPartitionLocation
+}
+
+func newObPartLocationEntry(partNum int) *ObPartLocationEntry {
+	entry := new(ObPartLocationEntry)
+	entry.partLocations = make(map[int64]*ObPartitionLocation, partNum)
+	return entry
+}
+
+func (e *ObPartLocationEntry) String() string {
+	var partitionLocationStr string
+	var i = 0
+	partitionLocationStr = partitionLocationStr + "{"
+	for k, v := range e.partLocations {
+		if i > 0 {
+			partitionLocationStr += ", "
+		}
+		i++
+		partitionLocationStr += "m[" + strconv.Itoa(int(k)) + "]="
+		if v != nil {
+			partitionLocationStr += v.String()
+		} else {
+			partitionLocationStr += "nil"
+		}
+	}
+	partitionLocationStr += "}"
+	return "ObPartLocationEntry{" +
+		"partLocations:" + partitionLocationStr +
+		"}"
+}
